Harvesters assignation to spawns/extensions
Harvesters now head for the closest spawn/extension whenever they want to unload. We keep track of how much energy will be after the harvester reaches its destination. For instance, if a structure has 150 energy with 200 capacity and we have two harvesters with 50 energy, only one of the harvester will be assigned to said structure; the other one will go somewhere else.